backend/requirements.txt
fastapi<=0.104.1
uvicorn<=0.24.0
pynetdicom<=2.0.2
pydicom<=2.4.3
websockets<=12.0
numpy<=1.26.2
Pillow<=10.1.0